pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "klarity"
version = "0.1.0"
description = "Uncertainty estimation for open-source generative models"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0.0,<2.6.0",
    "sentence-transformers>=3.0.0",
    "transformers>=4.41.0",
    "scikit-learn>=1.0.0",
    "numpy>=1.20.0",
    "scipy>=1.0.0",
    "pydantic>=2.0.0",
    "together==1.4.0",
    "matplotlib>=3.9.4",
    "ninja>=1.11.1.3",
    "cuda-python>=12.8.0",
    "nvidia-cuda-nvrtc-cu12>=12.4.127",
    "python-dotenv>=1.0.1",
    "vllm>=0.7.2",
    "setuptools>=75.8.0",
    "accelerate>=1.4.0",
    "xgrammar==0.1.11",
]

[project.urls]
Homepage = "https://github.com/klara-research/klarity"
Issues = "https://github.com/klara-research/klarity/issues"
Repository = "https://github.com/klara-research/klarity.git"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["klarity", "klarity.core"]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "ruff>=0.8.0"
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-s"
testpaths = ["tests"]

[tool.ruff]
line-length = 120
